+++ examples/ExampleReportsEvents.php
@@ -30,7 +30,7 @@
  * @author    Jeff Tanner <jefft@tune.com>
  * @copyright 2014 Tune (http://www.tune.com)
  * @license   http://opensource.org/licenses/MIT The MIT License (MIT)
- * @version   $Date: 2014-11-06 12:28:55 $
+ * @version   $Date: 2014-11-19 07:02:45 $
  * @link      https://developers.mobileapptracking.com @endlink
  *
  */
@@ -119,14 +119,14 @@ class ExampleReportsEvents
                 $response_timezone   = "America/Los_Angeles"
             );
 
-            echo "= TuneManagementResponse:" . PHP_EOL;
-            echo print_r($response, true) . PHP_EOL;
-
-            if ($response->getHttpCode() != 200) {
+            if (($response->getHttpCode() != 200) || ($response->getErrors() != null)) {
                 throw new \Exception(
-                    sprintf("Failed: %d: %s", $response->getHttpCode(), print_r($response->getErrors()))
+                    sprintf("Failed: %d: %s", $response->getHttpCode(), print_r($response, true))
                 );
             }
+
+            echo "= TuneManagementResponse:" . PHP_EOL;
+            echo print_r($response, true) . PHP_EOL;
 
             echo "= Count:" . $response->getData() . PHP_EOL;
 
@@ -136,22 +136,22 @@ class ExampleReportsEvents
             $response = $reports_logs_events->find(
                 $start_date,
                 $end_date,
-                $filter              = "(status = 'approved')",
                 $fields              = $reports_logs_events->fields(Events::TUNE_FIELDS_RECOMMENDED),
+                $filter              = "(status = 'approved')",
                 $limit               = 5,
                 $page                = null,
                 $sort                = array("created" => "DESC"),
                 $response_timezone   = "America/Los_Angeles"
             );
 
-            echo "= TuneManagementResponse:" . PHP_EOL;
-            echo print_r($response, true) . PHP_EOL;
-
-            if ($response->getHttpCode() != 200) {
+            if (($response->getHttpCode() != 200) || ($response->getErrors() != null)) {
                 throw new \Exception(
-                    sprintf("Failed: %d: %s", $response->getHttpCode(), print_r($response->getErrors()))
+                    sprintf("Failed: %d: %s", $response->getHttpCode(), print_r($response, true))
                 );
             }
+
+            echo "= TuneManagementResponse:" . PHP_EOL;
+            echo print_r($response, true) . PHP_EOL;
 
             echo "============================================================" . PHP_EOL;
             echo " Advertiser Logs Events CSV report for export.      " . PHP_EOL;
@@ -159,20 +159,20 @@ class ExampleReportsEvents
             $response = $reports_logs_events->export(
                 $start_date,
                 $end_date,
-                $filter              = "(status = 'approved')",
                 $fields              = $reports_logs_events->fields(Events::TUNE_FIELDS_RECOMMENDED),
+                $filter              = "(status = 'approved')",
                 $format              = "csv",
                 $response_timezone   = "America/Los_Angeles"
             );
 
-            echo "= TuneManagementResponse:" . PHP_EOL;
-            echo print_r($response, true) . PHP_EOL;
-
-            if ($response->getHttpCode() != 200) {
+            if (($response->getHttpCode() != 200) || ($response->getErrors() != null)) {
                 throw new \Exception(
-                    sprintf("Failed: %d: %s", $response->getHttpCode(), print_r($response->getErrors()))
+                    sprintf("Failed: %d: %s", $response->getHttpCode(), print_r($response, true))
                 );
             }
+
+            echo "= TuneManagementResponse:" . PHP_EOL;
+            echo print_r($response, true) . PHP_EOL;
 
             $job_id = Events::parseResponseReportJobId($response);
             echo "= CSV Job ID: {$job_id}" . PHP_EOL;
@@ -201,25 +201,25 @@ class ExampleReportsEvents
             $csv_report_reader->prettyPrint($limit = 5);
 
             echo "======================================================" . PHP_EOL;
-            echo " Advertiser Events JSON report for export.    " . PHP_EOL;
+            echo " Advertiser Events JSON report for export.            " . PHP_EOL;
             echo "======================================================" . PHP_EOL;
             $response = $reports_logs_events->export(
                 $start_date,
                 $end_date,
-                $filter              = "(status = 'approved')",
                 $fields              = $reports_logs_events->fields(Events::TUNE_FIELDS_RECOMMENDED),
+                $filter              = "(status = 'approved')",
                 $format              = "json",
                 $response_timezone   = "America/Los_Angeles"
             );
 
-            echo "= TuneManagementResponse:" . PHP_EOL;
-            echo print_r($response, true) . PHP_EOL;
-
-            if ($response->getHttpCode() != 200) {
+            if (($response->getHttpCode() != 200) || ($response->getErrors() != null)) {
                 throw new \Exception(
-                    sprintf("Failed: %d: %s", $response->getHttpCode(), print_r($response->getErrors()))
+                    sprintf("Failed: %d: %s", $response->getHttpCode(), print_r($response, true))
                 );
             }
+
+            echo "= TuneManagementResponse:" . PHP_EOL;
+            echo print_r($response, true) . PHP_EOL;
 
             $job_id = Events::parseResponseReportJobId($response);
             echo "= CSV Job ID: {$job_id}" . PHP_EOL;